Remove warnings for the net module.
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@1067 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/auth_cache_unittest.cc
@@ -65,7 +65,7 @@ TEST(AuthCacheTest, HttpKey) {
     "https://www.nowhere.org/WallyWorld"
   };
 
-  for (int i = 0; i < arraysize(url); i++) {
+  for (size_t i = 0; i < arraysize(url); i++) {
     std::string key = net::AuthCache::HttpKey(GURL(url[i]), *auth_info);
     EXPECT_EQ(expected[i], key);
   }
